% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCohortTable.R
\name{createCohortTable}
\alias{createCohortTable}
\title{createCohortTable function}
\usage{
createCohortTable(
  connection,
  oracleTempSchema,
  cohortDatabaseSchema,
  cohortTable_temp = "temp"
)
}
\arguments{
\item{T_genorateCohort}{}
}
\description{
This is a temp function which can create a sample Cohort Table
}
